VAI Implementation/requirements.txt
memory_profiler
scikit-image
pandas==1.0.5
numpy==1.17.2
scipy==1.3.1
ninja
torchvision==0.8.2
pillow
matplotlib
torch==1.7.1
torchaudio
torch-geometric
torchmetrics
scikit-learn
jupyter
notebook